Unify function to encode path
